worker/requirements.txt
numpy
pandas
werkzeug
pika
python-dotenv
requests